+++ go/CHANGELOG.md
@@ -1,3 +1,23 @@
 # Changelog
 
+## [0.2.0](https://github.com/launchdarkly/observability-sdk/compare/go/v0.1.0...go/v0.2.0) (2025-08-11)
+
+
+### Features
+
+* Add trace based sampling for go plugin. ([#170](https://github.com/launchdarkly/observability-sdk/issues/170)) ([dacbe76](https://github.com/launchdarkly/observability-sdk/commit/dacbe763d0ab672e1575290a37ad3f7cc2c3059a))
+
+## 0.1.0 (2025-08-07)
+
+
+### Features
+
+* Add export samplers. ([#143](https://github.com/launchdarkly/observability-sdk/issues/143)) ([76fd41e](https://github.com/launchdarkly/observability-sdk/commit/76fd41e24b6214b02e9380e184734d324fc59be4))
+* Add query for sampling config. ([#140](https://github.com/launchdarkly/observability-sdk/issues/140)) ([6800fb1](https://github.com/launchdarkly/observability-sdk/commit/6800fb109efee1ddbc9a2756c2b5c4c853bde4c4))
+* Basic go plugin implementation. ([#139](https://github.com/launchdarkly/observability-sdk/issues/139)) ([48badda](https://github.com/launchdarkly/observability-sdk/commit/48badda5a34f9cf6d510abb6277e2308d7ebbdd9))
+* Client-side event sampling custom sampler. ([#142](https://github.com/launchdarkly/observability-sdk/issues/142)) ([82df812](https://github.com/launchdarkly/observability-sdk/commit/82df812deacf626c8743043599ba584c79b059e1))
+* Sample out children of spans that have been sampled out. ([#155](https://github.com/launchdarkly/observability-sdk/issues/155)) ([a2d1d6e](https://github.com/launchdarkly/observability-sdk/commit/a2d1d6ef912e1befaf8fcad18291313942ef67f2))
+
+## Changelog
+
 All notable changes to the LaunchDarkly Observability SDK for Go will be documented in this file.