+++ sdk/@launchdarkly/observability-android/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Changelog
 
-## [0.2.0](https://github.com/launchdarkly/observability-sdk/compare/launchdarkly-observability-android-0.1.0...launchdarkly-observability-android-0.2.0) (2025-08-06)
+## [0.2.0](https://github.com/launchdarkly/observability-sdk/compare/launchdarkly-observability-android-0.2.0...launchdarkly-observability-android-0.2.0) (2025-08-11)
 
 
 ### Features
 
 * adds startSpan to Android observability sdk and evaluation events on spans ([#141](https://github.com/launchdarkly/observability-sdk/issues/141)) ([58b74b7](https://github.com/launchdarkly/observability-sdk/commit/58b74b727f54f58e94439df77740c12b10518a9c))
+
+
+### Miscellaneous Chores
+
+* fixing android package group for publishing ([#160](https://github.com/launchdarkly/observability-sdk/issues/160)) ([f3f4fc7](https://github.com/launchdarkly/observability-sdk/commit/f3f4fc729c7c29cbc0a6084b0f1cf352d9c6da39))
+* fixing android release ci caching ([#152](https://github.com/launchdarkly/observability-sdk/issues/152)) ([dd924eb](https://github.com/launchdarkly/observability-sdk/commit/dd924eb6330728c274d0bd99db6fcb0bc9b4ee7e))
+
+
+### Continuous Integration
+
+* adding group to observability-android gradle ([#156](https://github.com/launchdarkly/observability-sdk/issues/156)) ([c5a563c](https://github.com/launchdarkly/observability-sdk/commit/c5a563c9faf2048230b0682150ed1bca63506952))