+++ sdk/highlight-apollo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @highlight-run/apollo
 
+## 3.4.48
+
+### Patch Changes
+
+- @highlight-run/node@3.12.21
+
+## 3.4.47
+
+### Patch Changes
+
+- @highlight-run/node@3.12.20
+
 ## 3.4.46
 
 ### Patch Changes