+++ sdk/highlight-run/CHANGELOG.md
@@ -1,5 +1,17 @@
 # highlight.run
 
+## 9.19.1
+
+### Patch Changes
+
+- 1beae4d: add sampling support to react native o11y plugin
+
+## 9.19.0
+
+### Minor Changes
+
+- d181f89: Define session based on a sessionKey
+
 ## 9.18.23
 
 ### Patch Changes